pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "people-counter"
version = "0.1.0"
description = "Low-cost people counting for retail via stereo vision and edge AI"
requires-python = ">=3.11"
dependencies = [
    "opencv-contrib-python>=4.8",
    "numpy>=1.24",
    "paho-mqtt>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
hailo = ["hailo-platform>=4.17"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "black>=23.0", "ruff>=0.1"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
